SerializerRegistry is no longer public
